+++ src/cli/mod.rs
@@ -4,6 +4,7 @@ mod config;
 mod help;
 mod instance;
 mod internal;
+mod man;
 mod playlist_range;
 mod port;
 mod printers;
@@ -16,5 +17,6 @@ mod shell;
 //===========================================================================//
 
 pub use self::{
-    action::*, args::*, config::*, instance::*, props::*, run::*, shell::*,
+    action::*, args::*, config::*, instance::*, man::*, props::*, run::*,
+    shell::*,
 };